conda-requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
blas=1.0=openblas
bzip2=1.0.8=h5eee18b_6
ca-certificates=2024.11.26=h06a4308_0
expat=2.6.4=h6a678d5_0
freetype=2.12.1=h4a9f257_0
imagehash=4.3.1=pyhd8ed1ab_0
jpeg=9e=h5eee18b_3
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.40=h12ee557_0
lerc=3.0=h295c915_0
libdeflate=1.17=h5eee18b_1
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libopenblas=0.3.21=h043d6bf_0
libpng=1.6.39=h5eee18b_0
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.5.1=h6a678d5_0
libuuid=1.41.5=h5eee18b_0
libwebp-base=1.3.2=h5eee18b_1
lz4-c=1.9.4=h6a678d5_1
ncurses=6.4=h6a678d5_0
numpy=2.0.1=py312h2809609_1
numpy-base=2.0.1=py312he1a6c75_1
openjpeg=2.5.2=he7f1fd0_0
openssl=3.0.15=h5eee18b_0
pillow=11.0.0=py312hfdbf927_0
pip=24.2=py312h06a4308_0
python=3.12.8=h5148396_0
pywavelets=1.7.0=py312h5eee18b_0
readline=8.2=h5eee18b_0
scipy=1.14.1=py312h2809609_0
setuptools=75.1.0=py312h06a4308_0
sqlite=3.45.3=h5eee18b_0
tk=8.6.14=h39e8969_0
tzdata=2024b=h04d1e81_0
wheel=0.44.0=py312h06a4308_0
xz=5.4.6=h5eee18b_1
zlib=1.2.13=h5eee18b_1
zstd=1.5.6=hc292b87_0
